Expose NativeWindow.setThumbarButtons API to all platforms.
