+++ atom/browser/api/atom_api_window.h
@@ -11,12 +11,9 @@
 #include "base/memory/scoped_ptr.h"
 #include "ui/gfx/image/image.h"
 #include "atom/browser/api/trackable_object.h"
+#include "atom/browser/native_window.h"
 #include "atom/browser/native_window_observer.h"
 #include "native_mate/handle.h"
-
-#if defined(OS_WIN)
-#include "atom/browser/ui/win/thumbar_host.h"
-#endif
 
 class GURL;
 
@@ -133,10 +130,8 @@ class Window : public mate::TrackableObject<Window>,
   void SetProgressBar(double progress);
   void SetOverlayIcon(const gfx::Image& overlay,
                       const std::string& description);
-#if defined(OS_WIN)
   void SetThumbarButtons(
-      const std::vector<ThumbarHost::ThumbarButton>& buttons);
-#endif
+      const std::vector<NativeWindow::ThumbarButton>& buttons);
   void SetMenu(v8::Isolate* isolate, v8::Local<v8::Value> menu);
   void SetAutoHideMenuBar(bool auto_hide);
   bool IsMenuBarAutoHide();
